MINIFICPP-1755 - Separate span utils from gsl
