+++ libminifi/include/utils/gsl.h
@@ -27,28 +27,6 @@ namespace org::apache::nifi::minifi {
 namespace gsl = ::gsl_lite;
 
 namespace utils {
-namespace detail {
-template<typename T>
-using remove_cvref_t = typename std::remove_cv<typename std::remove_reference<T>::type>::type;
-}  // namespace detail
-
-template<typename Container, typename T>
-Container span_to(std::span<T> span) {
-  static_assert(std::is_constructible<Container, typename gsl::span<T>::iterator, typename gsl::span<T>::iterator>::value,
-      "The destination container must have an iterator (pointer) range constructor");
-  return Container(std::begin(span), std::end(span));
-}
-template<template<typename...> class Container, typename T>
-Container<detail::remove_cvref_t<T>> span_to(std::span<T> span) {
-  static_assert(std::is_constructible<Container<detail::remove_cvref_t<T>>, typename gsl::span<T>::iterator, typename gsl::span<T>::iterator>::value,
-      "The destination container must have an iterator (pointer) range constructor");
-  return span_to<Container<detail::remove_cvref_t<T>>>(span);
-}
-
-template<typename T, typename U>
-std::span<T> as_span(std::span<U> value) {
-  return std::span{reinterpret_cast<T*>(value.data()), value.size_bytes() / sizeof(T)};
-}
 
 template<typename T>
 struct is_not_null : std::false_type {};
@@ -56,6 +34,7 @@ template<typename T>
 struct is_not_null<gsl::not_null<T>> : std::true_type {};
 template<typename T>
 inline constexpr bool is_not_null_v = is_not_null<T>::value;
+
 }  // namespace utils
 
 }  // namespace org::apache::nifi::minifi
