+++ libminifi/test/unit/FlowFileSerializationTests.cpp
@@ -25,6 +25,7 @@
 #include "../TestBase.h"
 #include "../Catch.h"
 #include "utils/gsl.h"
+#include "utils/span.h"
 #include "FlowFile.h"
 #include "FlowFileRecord.h"
