+++ libminifi/test/unit/LineByLineInputOutputStreamCallbackTests.cpp
@@ -21,6 +21,7 @@
 #include "core/logging/LoggerConfiguration.h"
 #include "io/BufferStream.h"
 #include "spdlog/spdlog.h"
+#include "utils/span.h"
 
 using minifi::utils::LineByLineInputOutputStreamCallback;
